Platforms/TC2: move to new DtPlatformDxe driver
Replace the slightly overengineered FdtPlatformDxe driver with the new
and simple DtPlatformDxe driver, which simply exposes a single builtin
device tree binary to the OS.

Note that this driver is configurable, i.e., it allows ACPI to be
selected instead of device tree, in which case AcpiPlatformDxe is
allowed to run.

Note that this means we have to move the reference to FdtPlatformDxe.inf
from the shared .DSC to the individual .DSCs for Juno and FVP that will
keep using it for now.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>