Platforms/ARM/VExpress: Define extents of MM communication buffer
The secure world firmware (e.g. ARM Trusted Firmware) allocates the
communication buffer used in the MM Communicate SMC. EDK2 is expected to
map this buffer in its own address map. In the absence of a mechanism
where ARM TF can pass the extents of this buffer at runtime to EDK2, this
patch hard codes this information in the UEFI binary for the ARM FVP.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Achin Gupta <achin.gupta@arm.com>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>