Platforms/ARM/VExpress: Include MM communication protocol driver
This patch includes the DXE runtime driver that exports the
EFI_MM_COMMUNICATION_PROTOCOL for the FVP platform. This driver is
responsible for handling communication with the MM environment in the
secure world. An implementation of the ARM SVC lib is also include in the
FVP build.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Achin Gupta <achin.gupta@arm.com>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>