Platforms/FVP: Fix MADT to run the Foundation model in GICv3 mode with 4 CPUs
This upgrades the MADT table so that it exposes the GIC as a v3. The
virtual base address and interrupt, and the hypervisor base address are
corrected as well.

Since the Foundation model has 4 cores at the most, and since refactoring
this code to update the ACPI tables dynamically based on the actual core
count is more trouble that its worth, remove the second cluster while we
are at it.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Signed-off-by: James Morse <james.morse@arm.com>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>